backend/requirements.txt
numpy
google-genai
requests
